+++ ligo/skymap/tool/ligo_skymap_from_samples.py
@@ -59,6 +59,9 @@ def parser():
     parser.add_argument('--enable-multiresolution', action=EnableAction,
                         default=True,
                         help='generate a multiresolution HEALPix map')
+    parser.add_argument('--topnside', type=int, default=16,
+                        help='choose a start nside before HEALPix refinement '
+                        'steps (must be a valid nside)')
     parser.add_argument('-j', '--jobs', type=int, default=1, const=None,
                         nargs='?', help='Number of threads')
     parser.add_argument('--instruments', metavar='H1|L1|V1|...', nargs='+',
@@ -128,7 +131,7 @@ def main(args=None):
         skypost.jobs = args.jobs
 
     log.info('making skymap')
-    hpmap = skypost.as_healpix()
+    hpmap = skypost.as_healpix(top_nside=args.topnside)
     if not args.enable_multiresolution:
         hpmap = rasterize(hpmap)
     hpmap.meta.update(io.fits.metadata_for_version_module(version))