+++ CHANGES.rst
@@ -25,6 +25,10 @@ Changelog
 
 - Switch from WMAP9 to Planck15 cosmological parameters.
 
+- ``ligo.skymap.kde.Clustered2DSkyKDE.as_healpix()`` has an optional ``top_nside``
+  to allow for better initial grid, before refinement. 
+  ``ligo-skymap-from-samples`` has an additional ``--topnside`` argument, accordingly.
+
 0.1.12 (2019-09-19)
 ===================